refactor: fix type err
